// src/app/services/openapi-client/model/createCouponDto.ts
/**
 * ShopErpApi
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateCouponDto { 
    code: string;
    type: string;
    value: number;
    minPurchaseAmount?: number | null;
    maxDiscountAmount?: number | null;
    usageLimitTotal?: number | null;
    usageLimitPerUser?: number | null;
    validFrom?: string | null;
    validUntil?: string | null;
    isActive?: boolean;
}